move login from home component to login component
