+++ src/app/home/home.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
-import { LoginValidators } from './../validators/login.validators';
 
 @Component({
   selector: 'app-home',
@@ -10,23 +8,5 @@ import { LoginValidators } from './../validators/login.validators';
 export class HomeComponent implements OnInit {
   constructor() {}
 
-  loginForm = new FormGroup({
-    email: new FormControl('', [Validators.required, Validators.email]),
-    password: new FormControl('', [
-      Validators.required,
-      LoginValidators.shouldNotHaveSpaces
-    ])
-  });
-
-  get email() {
-    return this.loginForm.get('email');
-  }
-
-  get password() {
-    return this.loginForm.get('password');
-  }
-
   ngOnInit() {}
-
-  login() {}
 }
