+++ src/app/login/login.component.scss
@@ -1,0 +1,231 @@
+.main-bg {
+  background: #e5edf9;
+  min-height: 100vh;
+
+}
+
+h1 {
+  font-size: 2.8vw;
+
+  text-align: center;
+  padding: 1.5vw 1vw 2vw;
+  letter-spacing: 3px;
+  text-transform: uppercase;
+}
+
+.sub-main {
+  margin: 1.5vw 5vw;
+}
+
+.user-bg {
+  max-width: 400px;
+  margin: 0 auto;
+  background: #fff;
+  text-align: center;
+}
+
+.top-content {
+  padding: 2vw 4vw 4vw;
+  background: #1cc7d0;
+}
+
+.top-content img {
+  width: 100px;
+  height: 100px;
+  -webkit-border-radius: 50%;
+  -o-border-radius: 50%;
+  -ms-border-radius: 50%;
+  -moz-border-radius: 50%;
+  border-radius: 50%;
+}
+
+.sub-main form {
+  background: #ffff;
+  padding: 2em;
+  -webkit-box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+  -moz-box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+  box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+  margin: -2.5em 2.5em 2em;
+  -webkit-border-radius: 4px;
+  -o-border-radius: 4px;
+  -ms-border-radius: 4px;
+  -moz-border-radius: 4px;
+  border-radius: 4px;
+}
+
+p.legend {
+  color: #4e4d4d;
+  font-size: 24px;
+  text-align: center;
+  margin-bottom: 1.2em;
+}
+
+p.legend span {
+  color: #000;
+  margin-left: 10px;
+}
+
+.input {
+  position: relative;
+  margin: 20px auto;
+  width: 100%
+}
+
+.input span {
+  position: absolute;
+  display: block;
+  color: #1cc7d0;
+  left: 10px;
+  top: 12px;
+  font-size: 16px;
+}
+
+.input input {
+  width: 100%;
+  padding: 13px 10px 13px 34px;
+  display: block;
+  border: none;
+  border: 1px solid #1cc7d0;
+  color: #000;
+  box-sizing: border-box;
+  font-size: 13px;
+  outline: none;
+  letter-spacing: 1px;
+  background: #fff;
+  -webkit-box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+  -moz-box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+  box-shadow: 2px 5px 16px 2px rgba(16, 16, 16, 0.18);
+}
+
+.submit {
+  width: 105px;
+  height: 45px;
+  display: block;
+  margin: 2.5em auto 0;
+  background: #1cc7d0;
+  -webkit-border-radius: 10px;
+  -o-border-radius: 10px;
+  -ms-border-radius: 10px;
+  -moz-border-radius: 10px;
+  border-radius: 10px;
+  border: none;
+  cursor: pointer;
+  -webkit-transition: 0.5s all;
+  -o-transition: 0.5s all;
+  -moz-transition: 0.5s all;
+  -ms-transition: 0.5s all;
+  transition: 0.5s all;
+  color: white;
+}
+
+.button {
+  font-size: 1.3em;
+}
+
+.submit span {
+  color: #fff;
+  font-size: 20px;
+
+}
+
+.submit:hover {
+  opacity: .8;
+  -webkit-transition: 0.5s all;
+  -o-transition: 0.5s all;
+  -moz-transition: 0.5s all;
+  -ms-transition: 0.5s all;
+  transition: 0.5s all;
+}
+
+/* responsive */
+@media(max-width:1280px) {
+  .top-content {
+    padding: 3vw 4vw 5vw;
+  }
+}
+
+@media(max-width:1080px) {
+  h1 {
+    font-size: 3.5vw;
+  }
+}
+
+@media(max-width:991px) {
+  h1 {
+    font-size: 4vw;
+  }
+
+  .top-content {
+    padding: 3vw 4vw 6vw;
+  }
+}
+
+@media(max-width:800px) {
+  h1 {
+    font-size: 5vw;
+    padding: 2.5vw 1vw 3vw;
+  }
+
+  .top-content {
+    padding: 4vw 4vw 9vw;
+  }
+
+
+}
+
+@media(max-width:640px) {
+  h1 {
+    font-size: 6vw;
+    padding: 3vw 1vw 4vw;
+  }
+}
+
+
+
+@media(max-width:480px) {
+  h1 {
+    font-size: 2em;
+    letter-spacing: 1px;
+  }
+
+  .top-content {
+    padding: 2em 1em 4em;
+  }
+
+
+
+  p.legend {
+    font-size: 23px;
+  }
+
+  a.bottom-forgot {
+    font-size: 15px;
+  }
+}
+
+@media(max-width:384px) {
+  .sub-main form {
+    margin: -2.5em 2em 2em;
+  }
+
+  h1 {
+    padding: 5vw 1vw 6vw;
+  }
+}
+
+@media(max-width:320px) {
+  h1 {
+    font-size: 1.7em;
+  }
+
+  .sub-main form {
+    margin: -2em 1em 1.5em;
+    padding: 1.5em;
+  }
+
+  .top-content {
+    padding: 1.5em 1em 3em;
+  }
+
+
+}
