+++ src/app/login/login.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { LoginValidators } from './../validators/login.validators';
 
 @Component({
   selector: 'app-login',
@@ -6,10 +8,27 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./login.component.scss']
 })
 export class LoginComponent implements OnInit {
+  loading: false;
+  submitted: false;
+  loginForm: FormGroup;
 
-  constructor() { }
-
+  constructor() {}
   ngOnInit() {
+    this.loginForm = new FormGroup({
+      email: new FormControl('', [Validators.required, Validators.email]),
+      password: new FormControl('', [
+        Validators.required,
+        LoginValidators.shouldNotHaveSpaces
+      ])
+    });
+  }
+  get email() {
+    return this.loginForm.get('email');
   }
 
+  get password() {
+    return this.loginForm.get('password');
+  }
+
+  login() {}
 }
