--- src/app/signup/signup.component.html
@@ -1,3 +1,0 @@
-<p>
-  signup works!
-</p>
